+++ HISTORY.rst
@@ -16,6 +16,7 @@ Next Release
 ------------------
 
 * Add registrations limit to Wild Apricot events.
+* Work around imghdr bug of not recognizing some jpeg images.
 
 1.3.1 (2020-11-22)
 ------------------